Cloud sync update + bug fixes
- Cloud sync now support multiple OSs like it should have initially
- Added re-sync of remote data on game launching, along with general remote data notification
- Began moving all config things everywhere to globals in config.ts
- Fixed that annoying bug where the running game bar would shift into the app bar
- Added decompress to the compress.ts file

+++ src/plugins_bkg/cloud_saves.ts
@@ -1,20 +1,15 @@
 
 import { BrowserWindow, ipcMain, IpcMain } from "electron";
-import { compressFolder, CompressProgress } from "./tools/compression";
+import { compressFolder, CompressProgress, decompressFolder } from "./tools/compression";
 import { downloadFile, initWebDav, mutateFolder, webDavConfig } from "./nc_webdav";
 import { ErrorStats, Stats } from "node-downloader-helper";
 import { FileStat, WebDAVClient } from "webdav";
+import { getConfig, getOS, REMOTE_FILE_BASE, REMOTE_FOLDER } from "./config";
 import fs from "fs";
-import { getConfig } from "./config";
 import { Globals } from ".";
 import { GOG } from "@/types/gog/game_info";
 import os from "os";
-import zip from "node-stream-zip";
 
-// CONFIG ==>
-const REMOTE_FOLDER = ".game-saves";
-const REMOTE_FILE = "save.zip";
-// <== CONFIG
 
 let win = undefined as undefined | BrowserWindow;
 let globals = undefined as undefined | Globals;
@@ -23,6 +18,8 @@ function procSaveFile(save_file_raw: string, game: GOG.GameInfo){
   if(save_file_raw.startsWith("~")){
     return os.homedir()  + save_file_raw.substring(1);
   }else if(save_file_raw.startsWith("./")){
+    return game.root_dir  + save_file_raw.substring(1);
+  }else if(save_file_raw.startsWith("<steam>/")){
     return game.root_dir  + save_file_raw.substring(1);
   }
   return save_file_raw;
@@ -34,17 +31,92 @@ function getRemoteSaveDirectory(game: GOG.GameInfo){
   return mutated_folder + "/" + game.remote_name;
 }
 
+function getSavesLocation(game: GOG.GameInfo): undefined | GOG.GameSave{
+  const saves = game.remote?.saves as GOG.GameSavesLocation;
+  if(saves){
+    const os = getOS();
+    if(saves[os]){
+      const g_save =  (typeof saves[os] === "object" ? saves[os] : { "base": saves[os] }) as GOG.GameSave;
+      // Process the save paths
+      for(const key in g_save){
+        g_save[key] = procSaveFile(g_save[key], game);
+      }
+      return g_save;
+    }
+  }
+  return undefined;
+}
+
+function saveGamesExists(game: GOG.GameInfo, saves: GOG.GameSave): boolean{
+  for(const save in saves){
+    if(fs.existsSync(saves[save])){
+      return true;
+    }
+  }
+  return false;
+}
+
+async function packGameSave(game: GOG.GameInfo, saves: GOG.GameSave): Promise<string | undefined>{
+  const gog_path = getConfig("gog_path");
+  const tmp_dir = gog_path + "\\.temp\\";
+  const save_tmp = tmp_dir + "save_prep\\";
+  // Package each to a unique file, then compress them all together
+  fs.mkdirSync(save_tmp, { recursive: true });
+
+  let found = 0;
+  for(const save in saves){
+    if(fs.existsSync(saves[save])){
+      await compressFolder(saves[save], save_tmp + save + ".zip", (p: CompressProgress) =>{
+        win?.webContents.send("save-game-dl-progress", game, "Compressing save: " + save, p);
+      });
+      found++;
+    }
+  }
+  if(found <= 0){
+    return undefined;
+  }
+
+  // Compress final save folder
+  const output = tmp_dir + new Date().getTime() + "-save.zip";
+  await compressFolder(save_tmp, output, (p: CompressProgress) =>{
+    win?.webContents.send("save-game-dl-progress", game, "Compressing save: bundle", p);
+  }, 4);
+  fs.rmSync(save_tmp, { recursive: true });
+  return output;
+}
+
+async function unpackGameSave(game: GOG.GameInfo, saves: GOG.GameSave, save_package: string): Promise<void>{
+  const gog_path = getConfig("gog_path");
+  const tmp_dir = gog_path + "\\.temp\\";
+  const save_tmp = tmp_dir + "save_prep\\";
+  // Decompress master package
+  fs.mkdirSync(save_tmp, { recursive: true });
+  await decompressFolder(save_package, save_tmp, (p: CompressProgress) =>{
+    win?.webContents.send("save-game-dl-progress", game, "Unpacking save: bundle", p);
+  });
+
+  // Decompress save packages
+  for(const save in saves){
+    if(fs.existsSync(save_tmp + save + ".zip")){
+      await decompressFolder(save_tmp + save + ".zip", saves[save], (p: CompressProgress) =>{
+        win?.webContents.send("save-game-dl-progress", game, "Unpacking save: " + save, p);
+      });
+    }
+  }
+
+  fs.rmSync(save_tmp, { recursive: true });
+}
+
 export async function uploadGameSave(game: GOG.GameInfo){
   // Check if the cloud save location is specified
-  const save_file_raw = game?.remote?.save_location;
-  if(save_file_raw === undefined){
+  const save_files = getSavesLocation(game);
+  if(save_files === undefined){
     return;
   }
-  const local_save_folder = procSaveFile(save_file_raw, game);
-  if(!fs.existsSync(local_save_folder)){
+  if(!saveGamesExists(game, save_files)){
     globals?.notify({
       title: "Cloud Save",
-      text: "Failed to locate saves for game " + game.remote_name + " in folder: " + local_save_folder,
+      text: "Failed to locate saves for game " + game.remote_name,
       type: "warning",
       sticky: true
     });
@@ -58,20 +130,22 @@ export async function uploadGameSave(game: GOG.GameInfo){
 
   if(nc_cfg !== undefined && web_dav !== undefined){
     const remote_save_folder = getRemoteSaveDirectory(game);
-    const remove_save_file = remote_save_folder + "/" + REMOTE_FILE;
+    const remove_save_file = remote_save_folder + "/" + REMOTE_FILE_BASE + ".zip";
     // Make the game folder here
     if(!await web_dav.exists(remote_save_folder)){
       await web_dav.createDirectory(remote_save_folder, {recursive: true});
     }
-    const gog_path = getConfig("gog_path");
-    const tmp_dir = gog_path + "\\.temp\\";
-    // Compress the save file
-    const save_zip = tmp_dir + new Date().getTime() + "-save.zip";
-    win?.webContents.send("save-game-dl-progress", game, "Compressing save", {total: 100, progress: 0});
-    // ========== FILE ZIPPING ==========
-    await compressFolder(local_save_folder, save_zip, (p: CompressProgress) =>{
-      win?.webContents.send("save-game-dl-progress", game, "Compressing save", p);
-    });
+
+    const save_zip = await packGameSave(game, save_files);
+    if(save_zip === undefined){
+      globals?.notify({
+        title: "Cloud Save",
+        text: "Failed to package saves for " + game.remote_name,
+        type: "error",
+        sticky: true
+      });
+      return;
+    }
 
     // Time to upload!!
     win?.webContents.send("save-game-dl-progress", game, "Uploading save", {total: -1, progress: -1});
@@ -87,11 +161,13 @@ export async function uploadGameSave(game: GOG.GameInfo){
           remove_save_file
             + "." + d.getFullYear() + "-" + d.getMonth() + "-" + d.getDate()
             + "." + d.getHours() + "-" + d.getMinutes()
-            + ".old");
+            + ".zip");
       }
       await web_dav.moveFile(remove_save_file + ".new",  remove_save_file);
       // Update folder time
-      fs.utimesSync(local_save_folder, new Date(), new Date());
+      for(const save in save_files){
+        fs.utimesSync(save_files[save], new Date(), new Date());
+      }
     }
     read_stream.close();
     // Delete tmp save file
@@ -109,12 +185,17 @@ async function deployCloudSave(
   game: GOG.GameInfo,
   web_dav: WebDAVClient,
   remove_save_file: string,
-  local_save_folder: string,
   resolver: (cloud: boolean) => void){
+
+  const save_files = getSavesLocation(game);
+  if(save_files === undefined){
+    return;
+  }
+
   win?.webContents.send("save-game-dl-progress", game, "Downloading", {total: 100, progress: 0});
   const gog_path = getConfig("gog_path");
   const tmp_download = gog_path + "\\.temp\\";
-  const save_download = game.name + "-save.zip";
+  const save_download = game.remote_name + "-save.zip";
   if(fs.existsSync(tmp_download + "/" + save_download)){
     fs.rmSync(tmp_download + "/" + save_download);
   }
@@ -127,27 +208,13 @@ async function deployCloudSave(
       type: "info"
     });
     win?.webContents.send("save-game-sync-state", game, "Unpacking");
-    const archive = new zip.async({file: tmp_download + save_download});
-    const total_count = await archive.entriesCount;
-    let ex_count = 0;
-    win?.webContents.send("save-game-dl-progress", game, "Unpacking", {total: total_count, progress: ex_count});
-    archive.on("extract", () => {
-      ex_count++;
-      win?.webContents.send("save-game-dl-progress", game, "Unpacking", {total: total_count, progress: ex_count});
-    });
-    // Remove old save folder
-    if(fs.existsSync(local_save_folder)){
-      fs.rmSync(local_save_folder, {recursive: true});
-    }
-    const count = await archive.extract(null, local_save_folder);
-    console.log("Extracted:" + count);
+    await unpackGameSave(game, save_files, tmp_download + save_download);
     win?.webContents.send("save-game-stopped", game);
     globals?.notify({
       title: "Cloud Save",
       text: "Saves for game " + game.remote_name + " synchronized locally",
       type: "success"
     });
-    await archive.close();
     resolver(true);
   });
   active_dl.on("stop", () => {
@@ -169,10 +236,38 @@ async function deployCloudSave(
   active_dl.start();
 }
 
+async function newerInCloud(
+  game: GOG.GameInfo,
+  web_dav: WebDAVClient,
+  remote_save_folder: string,
+  remote_save_file: string
+): Promise<boolean>{
+  const save_files = getSavesLocation(game);
+  if(save_files === undefined){
+    return false;
+  }
+  // Check each folder for earliest date
+  let oldest = -1;
+  for(const s in save_files){
+    const f = fs.statSync(save_files[s]);
+    if(oldest === -1 || oldest > f.mtimeMs){
+      oldest = f.mtimeMs;
+    }
+  }
+  // If the folder or file doesn't exist remote there is clearly no newer saves
+  if(!(await web_dav.exists(remote_save_folder)) || !(await web_dav.exists(remote_save_file))){
+    return false;
+  }
+  // Get save age in cloud
+  const stat = await web_dav.stat(remote_save_file) as FileStat;
+  console.log("oldest:", oldest, "Date.parse(stat.lastmod):", Date.parse(stat.lastmod));
+  return oldest < Date.parse(stat.lastmod);
+}
+
 export async function syncGameSave(game: GOG.GameInfo, resolver: (cloud: boolean) => void){
   // Check if the cloud save location is specified
-  const save_file_raw = game?.remote?.save_location;
-  if(save_file_raw === undefined){
+  const save_files = getSavesLocation(game);
+  if(save_files === undefined){
     console.warn("Failed to find save game location for game: ", game);
     return resolver(false);
   }
@@ -181,54 +276,61 @@ export async function syncGameSave(game: GOG.GameInfo, resolver: (cloud: boolean
 
   if(nc_cfg !== undefined && web_dav !== undefined){
     win?.webContents.send("save-game-sync-start", game);
+    // Check if newer in cloud
+    win?.webContents.send("save-game-sync-search", game);
+    // Generate cloud location
     const remote_save_folder = getRemoteSaveDirectory(game);
-    console.log("remote_save_folder", remote_save_folder);
-    const local_save_folder = procSaveFile(save_file_raw, game);
-    const remove_save_file = remote_save_folder + "/" + REMOTE_FILE;
-    // Make the game folder here
-    if(!(await web_dav.exists(remote_save_folder))){
+    const remote_save_file = remote_save_folder + "/" + REMOTE_FILE_BASE + ".zip";
+    if(!(await newerInCloud(game, web_dav, remote_save_folder, remote_save_file))){
       win?.webContents.send("save-game-stopped", game);
       return resolver(false);
     }
 
-    win?.webContents.send("save-game-sync-search", game);
-    console.log("Looking for file: ", remove_save_file);
-    if(await web_dav.exists(remove_save_file)){
-      // Check if it is newer than the local files
-      const local_stat = fs.statSync(local_save_folder);
-      const remote_stat = await web_dav.stat(remove_save_file) as FileStat;
-      if((local_stat.mtimeMs + 10000) < Date.parse(remote_stat.lastmod)){
-        const evt = "use-cloud-save-" + new Date().getTime();
-        const evtl = "use-local-save-" + new Date().getTime();
-        globals?.notify({
-          title: "Cloud Save",
-          text: "Newer saves for game " + game.remote_name + " in cloud",
-          type: "info",
-          actions: [
-            {
-              name: "Use cloud save",
-              event: evt,
-              clear: true
-            },
-            {
-              name: "Use local save",
-              event: evtl,
-              clear: true
-            }
-          ],
-          sticky: true
-        });
-        win?.webContents.send("save-game-stopped", game);
-        ipcMain.once(evt, () => {
-          deployCloudSave(game, web_dav, remove_save_file, local_save_folder, resolver);
-        });
-        ipcMain.once(evtl, () => {
-          resolver(false);
-        });
-        return;
+    // At this point we know the file exists, and is newer than local saves
+
+    // Proceed to user request
+    const evt = "use-cloud-save-" + new Date().getTime();
+    const evtl = "use-local-save-" + new Date().getTime();
+    globals?.notify({
+      title: "Cloud Save",
+      text: "Newer saves for game " + game.remote_name + " in cloud",
+      type: "info",
+      actions: [
+        {
+          name: "Use cloud save",
+          event: evt,
+          clear: true
+        },
+        {
+          name: "Use local save",
+          event: evtl,
+          clear: true
+        }
+      ],
+      sticky: true
+    });
+    win?.webContents.send("save-game-stopped", game);
+    const fn = {
+      deploy: ()=> {
+        // Nothing
+      },
+      cancel: ()=> {
+        // Nothing
       }
-      win?.webContents.send("save-game-stopped", game);
-    }
+    };
+    const deploy = () => {
+      deployCloudSave(game, web_dav, remote_save_file, resolver);
+      ipcMain.off(evtl, fn.cancel);
+    };
+    fn.deploy = deploy;
+    const cancel = () => {
+      resolver(false);
+      ipcMain.off(evt, fn.deploy);
+    };
+    fn.cancel = cancel;
+    ipcMain.once(evt, deploy);
+    ipcMain.once(evtl, cancel);
+    return;
   }else{
     console.error("Failed to get web dav connection for save sync!");
   }
